Make broadcast_local_sig function share among message prcoessor functions
